build: cargo multi project
