+++ Cargo.toml
@@ -1,17 +1,7 @@
-[package]
-name = "vpinball-plugin-rust"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "vpinball_plugin_rust"
-crate-type = ["lib", "cdylib", "staticlib"]
-
-[dependencies]
-log = "0.4.22"
-simple_logger = "5.0.0"
-
-[build-dependencies]
-bindgen = "0.70.1"
-# for downloading the vpinball plugin header
-reqwest = { version = "0.12", features = ["blocking"] }
+[workspace]
+members = [
+    "plugin",
+    "fpscounter",
+    "rainbow",
+]
+resolver = "2"
