+++ plugin/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "vpinball-plugin-api"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+log = "0.4.22"
+#simple_logger = "5.0.0"
+
+[build-dependencies]
+bindgen = "0.70.1"
+# for downloading the vpinball plugin header
+reqwest = { version = "0.12", features = ["blocking"] }
